Feature [#21] Create Primary Key for Contracts Table (#22)
* feat: create primary key for contracts table

* fix: pharmacy_id works not compatible to phar_id

+++ database/migrations/2024_09_17_111441_create_works_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('works', function (Blueprint $table) {
             $table->unsignedBigInteger('employee_id');
-            $table->unsignedInteger('pharmacy_id');
+            $table->unsignedBigInteger('pharmacy_id');
             $table->datetime('shift_start');
             $table->datetime('shift_end');
             $table->timestamps();
